+++ insidapy/simulate/utils/ode_default_solve_call.py
@@ -1,6 +1,6 @@
 
 import numpy as np
-from scipy.integrate import odeint
+from scipy.integrate import solve_ivp
 
 from .ode_collection import *     # Imports all the ODEs from the collection file.
 
@@ -19,23 +19,29 @@ def solve_ode(self, y0, tspan):
 
     # ..........................................
     if self.example == 'batch1':
-        y = odeint(func=batch1, y0=y0, t=tspan)
+        y = solve_ivp(fun=batch1, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
     # ..........................................
     elif self.example == 'batch2':
-        y = odeint(func=batch2, y0=y0, t=tspan)
+        y = solve_ivp(fun=batch2, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
     # ..........................................
     elif self.example == 'batch3':
-        y = odeint(func=batch3, y0=y0, t=tspan)
+        y = solve_ivp(fun=batch3, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
     # ..........................................
     elif self.example == 'batch4':
-        y = odeint(func=batch4, y0=y0, t=tspan)
+        y = solve_ivp(fun=batch4, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
     # ..........................................
     elif self.example == 'batch5':
-        y = odeint(func=batch5, y0=y0, t=tspan)
+        y = solve_ivp(fun=batch5, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
+    # ..........................................
+    elif self.example == 'batch6':
+        y = solve_ivp(fun=batch6, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
         
     # ..........................................
     elif self.example == 'fedbatch1':
-        y = odeint(func=fedbatch1, y0=y0, t=tspan)
+        y = solve_ivp(fun=fedbatch1, y0=y0, t_span=[tspan[0], tspan[-1]], t_eval=tspan)
         
-        
+    # ..........................................
+    # Get only species data
+    # ..........................................
+    y = y.y.T
     return y